Move GeneralPredicates logic to kubelet.
